start auth feature

+++ rubyTraining/app/graphql/types/mutation_type.rb
@@ -2,6 +2,7 @@
 
 module Types
   class MutationType < Types::BaseObject
+    require_dependency 'types/user_type'
     # TODO: remove me
 #     field :test_field, String, null: false,
 #       description: "An example field added by the generator"
@@ -50,7 +51,17 @@ module Types
     end
 
     def create_article(title:, body:, author_id:, published_at: nil)
-      Article.create(title: title, body: body, author_id: author_id, published_at: published_at)
+      authenticate_user!
+      article = Article.new(
+        title: title,
+        body: body,
+        author_id: author_id,
+        published_at: published_at,
+        user: context[:current_user]
+      )
+      authorize article, :create?
+      article.save!
+      article
     end
 
     field :update_article, ArticleType, null: false do
@@ -62,8 +73,15 @@ module Types
     end
 
     def update_article(id:, title: nil, body: nil, author_id: nil, published_at: nil)
+      authenticate_user!
       article = Article.find(id)
-      article.update(title: title, body: body, author_id: author_id, published_at: published_at)
+      authorize article, :update?
+      article.update!(
+        title: title,
+        body: body,
+        author_id: author_id,
+        published_at: published_at
+      )
       article
     end
 
@@ -72,9 +90,54 @@ module Types
     end
 
     def delete_article(id:)
+      authenticate_user!
       article = Article.find(id)
+      authorize article, :destroy?
       article.destroy
       true
+    end
+
+    # User mutations
+    field :sign_up, UserType, null: false do
+      argument :name, String, required: true
+      argument :email, String, required: true
+      argument :password, String, required: true
+    end
+
+    def sign_up(name:, email:, password:)
+      user = User.create(name: name, email: email, password: password)
+      if user.persisted?
+        user
+      else
+        raise GraphQL::ExecutionError, "Failed to sign up: #{user.errors.full_messages.join(', ')}"
+      end
+    end
+
+    field :log_in, String, null: false do
+      argument :email, String, required: true
+      argument :password, String, required: true
+    end
+
+    def log_in(email:, password:)
+      user = User.find_for_authentication(email: email)
+      if user&.valid_password?(password)
+        JsonWebToken.encode(user_id: user.id)
+      else
+        raise GraphQL::ExecutionError, 'Invalid email or password'
+      end
+    end
+
+    private
+
+    def authenticate_user!
+      raise GraphQL::ExecutionError, "You need to log in to perform this action" unless context[:current_user]
+    end
+
+    def authorize(record, query)
+      policy = Pundit.policy!(context[:current_user], record)
+      unless policy.public_send(query)
+        raise GraphQL::ExecutionError, "You are not authorized to perform this action"
+      end
     end
   end
 end
